Split out common redirects and vanity middleware
