+++ modules/__tests__/urls.test.js
@@ -2,6 +2,7 @@
 'use strict';
 
 const {
+    cleanLinkNoise,
     getBaseUrl,
     getCurrentUrl,
     hasTrailingSlash,
@@ -15,6 +16,24 @@ const {
 const httpMocks = require('node-mocks-http');
 
 describe('URL Helpers', () => {
+    describe('#cleanLinkNoise', () => {
+        it('should clean link noise from the URL', () => {
+            expect(cleanLinkNoise('/funding/programmes/reaching-communities-england')).toBe(
+                '/funding/programmes/reaching-communities-england'
+            );
+
+            expect(cleanLinkNoise('/~/link.aspx')).toBe('/');
+
+            expect(cleanLinkNoise('/research/health-and-well-being/~/~/~/~/~/~/~/~/~/~/~/~/~/~/~/~/link.aspx')).toBe(
+                '/research/health-and-well-being/'
+            );
+
+            expect(cleanLinkNoise('/welsh/england/funding/funding-guidance/applying-for-funding/~/link.aspx')).toBe(
+                '/welsh/england/funding/funding-guidance/applying-for-funding/'
+            );
+        });
+    });
+
     describe('#getCurrentUrl', () => {
         it('should return expected url for en locale', () => {
             const req = httpMocks.createRequest({
